Backend/requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2
django-environ==0.4.5
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==7.0.0
mysqlclient==1.4.2.post1
pluggy==0.9.0
py==1.8.0
PyJWT==1.7.1
pylint==2.3.1
pytest==4.4.0
pytest-django==3.4.8
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
wrapt==1.11.1
